+++ sdk/cognitiveservices/azure-mgmt-cognitiveservices/generated_tests/test_cognitive_services_management_accounts_operations.py
@@ -102,7 +102,9 @@ class TestCognitiveServicesManagementAccountsOperations(AzureMgmtRecordedTestCas
                             {"id": "str", "ignoreMissingVnetServiceEndpoint": bool, "state": "str"}
                         ],
                     },
-                    "networkInjections": {"scenario": "str", "subnetArmId": "str", "useMicrosoftManagedNetwork": bool},
+                    "networkInjections": [
+                        {"scenario": "str", "subnetArmId": "str", "useMicrosoftManagedNetwork": bool}
+                    ],
                     "privateEndpointConnections": [
                         {
                             "etag": "str",
@@ -259,7 +261,9 @@ class TestCognitiveServicesManagementAccountsOperations(AzureMgmtRecordedTestCas
                             {"id": "str", "ignoreMissingVnetServiceEndpoint": bool, "state": "str"}
                         ],
                     },
-                    "networkInjections": {"scenario": "str", "subnetArmId": "str", "useMicrosoftManagedNetwork": bool},
+                    "networkInjections": [
+                        {"scenario": "str", "subnetArmId": "str", "useMicrosoftManagedNetwork": bool}
+                    ],
                     "privateEndpointConnections": [
                         {
                             "etag": "str",